Display time on conversations list screen
Moved conversation displayed in conversations list screen to its own
component.
It allows to have a computed property to display time
